Fix shader patterns that relied on canvas compositing
In WebGL, alpha values from any fragments (even those with blending off) are used to composite the color buffer with the canvas background.
This does not happen in OpenGL ES. The 'stripes' pattern works in WebGL by changing the alpha value of a fragment and allowing the canvas to
blend the background color into the final result. In OpenGL ES, no such blending happens. This change applies the fragment alpha to the color
values before canvas composition and then returns the alpha value to 1, resulting in consistent behavior in WebGL and OpenGL ES.